pyproject.toml
[tool.poetry]
name = "blog-db"
version = "0.1.0"
description = ""
authors = ["SuperMuel <69467005+SuperMuel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.7.3"
motor = "^3.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
